Updating WebServer Console with AppFunc
Added the AppFunc delegate with a signature  of IDictionary<string, object> returning a Task. This allows the OWIN pipeline to process component after component by passing it to the next AppFunc variable.

Added a reference to the current environment of owin.ResponseBody in Invoke(). This allowed us to write a message to send to the user with every request made.