Update WebServer Console with Middleware
Added middleware located in the Configuration method of the StartUp class. Allows us to see how middleware functions with OWIN components through the pipeline. 

In the Use() method of both middleware we used an async lambda expression to process the environment and await the next component coming through the pipeline. With each request we wrote the console some data from the environment dictionary.